Fix in SpecialKeys fix in graphics cpu usage better seeking and time duration estimation. fix in fxb,fxp saving. other smaller changes.

+++ trunk/psycle/src/psycle/host/SpecialKeys.hpp
@@ -13,8 +13,8 @@ public:
 	CSpecialKeys(CWnd* pParent = NULL);   // standard constructor
 	enum { IDD = IDD_SPECIALKEYS };
 
-	int key;
-	int mod;
+	WORD key;
+	WORD mod;
 
 protected:
 	virtual void DoDataExchange(CDataExchange* pDX);    // DDX/DDV support
